fixed bug in the main branch

+++ src/main/java/com/example/ElectricityTokenGenerator/dto/Tokens/TokensGenerationDTO.java
@@ -16,5 +16,5 @@ public class TokensGenerationDTO {
     private String serialNumber;
     private LocalDateTime createdAt;
     private LocalDateTime expiredAt;
-    private Long kiloWatts;
+    private Double kiloWatts;
 }
